Completed AI Project Model
Over the past few days, we have both been working collaboratively on this code over Discord. We are now seeing good outputs for all 3 models.
Changes so far have not been pushed due to the fact we have been working on it together on a local copy on one of our machines, while sharing screens.